# cmd/api/admin-api/docs/swagger.yaml
basePath: /
definitions:
  apiresp.ApiResponse:
    properties:
      data: {}
      errCode:
        type: integer
      errDlt:
        type: string
      errMsg:
        type: string
    type: object
...
  github_com_openimsdk_chat_pkg_common_apistruct.NewUserCountResp:
    properties:
      date_count:
        additionalProperties:
          type: integer
        type: object
      total:
        type: integer
    type: object
...
  github_com_openimsdk_chat_pkg_protocol_admin.AddIPForbiddenReq:
    properties:
      forbiddens:
        items:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.IPForbiddenAdd'
        type: array
    type: object
...
  github_com_openimsdk_chat_pkg_protocol_admin.AddUserIPLimitLoginReq:
    properties:
      limits:
        items:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.UserIPLimitLogin'
        type: array
    type: object
...
  github_com_openimsdk_chat_pkg_protocol_admin.DelIPForbiddenReq:
    properties:
      ips:
        items:
          type: string
        type: array
    type: object
...
  github_com_openimsdk_chat_pkg_protocol_admin.DelUserIPLimitLoginReq:
    properties:
      limits:
        items:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.UserIPLimitLogin'
        type: array
    type: object
...
  github_com_openimsdk_chat_pkg_protocol_admin.IPForbidden:
    properties:
      createTime:
        type: integer
      ip:
        type: string
      limitLogin:
        type: boolean
      limitRegister:
        type: boolean
    type: object
  github_com_openimsdk_chat_pkg_protocol_admin.IPForbiddenAdd:
    properties:
      ip:
        type: string
      limitLogin:
        type: boolean
      limitRegister:
        type: boolean
    type: object
...
  github_com_openimsdk_chat_pkg_protocol_admin.LimitUserLoginIP:
    properties:
      createTime:
        type: integer
      ip:
        type: string
      user:
        $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_common.UserPublicInfo'
      userID:
        type: string
    type: object
...
  github_com_openimsdk_chat_pkg_protocol_admin.SearchIPForbiddenReq:
    properties:
      keyword:
        type: string
      pagination:
        $ref: '#/definitions/sdkws.RequestPagination'
      status:
        type: integer
    type: object
  github_com_openimsdk_chat_pkg_protocol_admin.SearchIPForbiddenResp:
    properties:
      forbiddens:
        items:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.IPForbidden'
        type: array
      total:
        type: integer
    type: object
...
  github_com_openimsdk_chat_pkg_protocol_admin.SearchUserIPLimitLoginReq:
    properties:
      keyword:
        type: string
      pagination:
        $ref: '#/definitions/sdkws.RequestPagination'
    type: object
  github_com_openimsdk_chat_pkg_protocol_admin.SearchUserIPLimitLoginResp:
    properties:
      limits:
        items:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.LimitUserLoginIP'
        type: array
      total:
        type: integer
    type: object
...
  github_com_openimsdk_chat_pkg_protocol_admin.UserIPLimitLogin:
    properties:
      ip:
        type: string
      userID:
        type: string
    type: object
  github_com_openimsdk_chat_pkg_protocol_chat.AddUserAccountReq:
    properties:
      deviceID:
        type: string
      ip:
        type: string
      platform:
        type: integer
      user:
        $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_chat.RegisterUserInfo'
    type: object
  github_com_openimsdk_chat_pkg_protocol_chat.ChangePasswordReq:
    properties:
      currentPassword:
        type: string
      newPassword:
        type: string
      userID:
        type: string
    type: object
...
  github_com_openimsdk_chat_pkg_protocol_chat.RegisterUserInfo:
    properties:
      RegisterType:
        type: integer
      account:
        type: string
      areaCode:
        type: string
      birth:
        type: integer
      email:
        type: string
      faceURL:
        type: string
      gender:
        type: integer
      nickname:
        type: string
      password:
        type: string
      phoneNumber:
        type: string
      userID:
        type: string
    type: object
  github_com_openimsdk_chat_pkg_protocol_chat.SetAllowRegisterReq:
    properties:
      allowRegister:
        type: boolean
    type: object
  github_com_openimsdk_chat_pkg_protocol_chat.UserLoginCountReq:
    properties:
      end:
        type: integer
      start:
        type: integer
    type: object
  github_com_openimsdk_chat_pkg_protocol_chat.UserLoginCountResp:
    properties:
      count:
        additionalProperties:
          type: integer
        type: object
      loginCount:
        type: integer
      unloginCount:
        type: integer
    type: object
...
  github_com_openimsdk_chat_pkg_protocol_common.UserPublicInfo:
    properties:
      account:
        type: string
      email:
        type: string
      faceURL:
        type: string
      gender:
        type: integer
      level:
        type: integer
      nickname:
        type: string
      userID:
        type: string
    type: object
  sdkws.RequestPagination:
    properties:
      pageNumber:
        type: integer
      showNumber:
        type: integer
    type: object
  user.UserRegisterCountReq:
    properties:
      end:
        type: integer
      start:
        type: integer
    type: object
...
host: localhost:10009
info:
  contact:
    email: ilovecoding@foxmail.com
    name: 飞宏IM技术支持
    url: https://github.com/liony823/chat
  description: 飞宏IM后台管理接口
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: 飞宏IM后台管理接口
  version: 1.0.0
paths:
...
  /account/add_user:
    post:
      consumes:
      - application/json
      description: 创建新的普通用户账户
      operationId: addUserAccount
      parameters:
      - description: 用户账户信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_chat.AddUserAccountReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiresp.ApiResponse'
      summary: 添加用户账户
      tags:
      - user
...
  /forbidden/ip/add:
    post:
      consumes:
      - application/json
      description: 添加禁止注册和登录的IP
      operationId: addIPForbidden
      parameters:
      - description: IP黑名单信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.AddIPForbiddenReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiresp.ApiResponse'
      summary: 添加IP黑名单
      tags:
      - ip_limit
  /forbidden/ip/del:
    post:
      consumes:
      - application/json
      description: 删除禁止注册和登录的IP
      operationId: delIPForbidden
      parameters:
      - description: 要删除的IP信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.DelIPForbiddenReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiresp.ApiResponse'
      summary: 删除IP黑名单
      tags:
      - ip_limit
  /forbidden/ip/search:
    post:
      consumes:
      - application/json
      description: 查询被禁止注册和登录的IP列表
      operationId: searchIPForbidden
      parameters:
      - description: 搜索条件
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.SearchIPForbiddenReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apiresp.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.SearchIPForbiddenResp'
              type: object
      summary: 搜索IP黑名单
      tags:
      - ip_limit
  /forbidden/user/add:
    post:
      consumes:
      - application/json
      description: 限制用户在指定IP地址登录
      operationId: addUserIPLimitLogin
      parameters:
      - description: IP限制信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.AddUserIPLimitLoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiresp.ApiResponse'
      summary: 添加用户IP登录限制
      tags:
      - forbidden
  /forbidden/user/del:
    post:
      consumes:
      - application/json
      description: 删除指定的用户IP登录限制
      operationId: delUserIPLimitLogin
      parameters:
      - description: 要删除的限制信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.DelUserIPLimitLoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiresp.ApiResponse'
      summary: 删除用户IP登录限制
      tags:
      - ip_limit
  /forbidden/user/search:
    post:
      consumes:
      - application/json
      description: 查询用户IP登录限制列表
      operationId: searchUserIPLimitLogin
      parameters:
      - description: 搜索条件
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.SearchUserIPLimitLoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apiresp.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_admin.SearchUserIPLimitLoginResp'
              type: object
      summary: 搜索用户IP登录限制
      tags:
      - ip_limit
...
  /statistic/login_user_count:
    post:
      consumes:
      - application/json
      description: 获取用户登录统计数据
      operationId: loginUserCount
      parameters:
      - description: 统计条件
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_chat.UserLoginCountReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apiresp.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_chat.UserLoginCountResp'
              type: object
      summary: 统计用户登录数
      tags:
      - statistic
  /statistic/new_user_count:
    post:
      consumes:
      - application/json
      description: 获取新注册用户统计数据
      operationId: newUserCount
      parameters:
      - description: 统计条件
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/user.UserRegisterCountReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apiresp.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_openimsdk_chat_pkg_common_apistruct.NewUserCountResp'
              type: object
      summary: 统计新增用户数
      tags:
      - statistic
...
  /user/allow_register/set:
    post:
      consumes:
      - application/json
      description: 设置系统是否允许新用户注册
      operationId: setAllowRegister
      parameters:
      - description: 注册配置
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_chat.SetAllowRegisterReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiresp.ApiResponse'
      summary: 设置是否允许注册
      tags:
      - user
...
  /user/password/reset:
    post:
      consumes:
      - application/json
      description: 重置指定用户的密码
      operationId: resetUserPassword
      parameters:
      - description: 密码重置信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_openimsdk_chat_pkg_protocol_chat.ChangePasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiresp.ApiResponse'
      summary: 重置用户密码
      tags:
      - user
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: '"Type ''Bearer'' followed by a space and JWT token"'
    in: header
    name: token
    type: apiKey
swagger: "2.0"